pyproject.toml
[tool.poetry]
name = "rabbitmq-olm-package-repo"
version = "0.1.0"
description = "Script to generate OLM bundles for RabbitMQ operators"
authors = ["RabbitMQ Platform team"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
isort = "^5.9.3"
python = "^3.12"
PyYAML = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
generate_bundle = "rabbitmq_olm_package_repo:main"
